requirements.txt
django
djangorestframework
psycopg2-binary==2.9.6
redis==4.6.0
django-rq==2.8.1